+++ snippets/panelBar.php
@@ -1,13 +1,7 @@
 
-<?php
-$panelBar = new Kirby\panelBar\Core([
-  'elements' => isset($elements) ? $elements : null,
-]);
-?>
+<?php $panelBar = new Kirby\panelBar\Core(); ?>
 
-
-<?php if($panelBar->isShown() and !get('hidePanelBar')) : ?>
-
+<?php if($user = site()->user() and $user->hasPanelAccess()) : ?>
   <div class="panelBar panelBar-reset <?= $panelBar->classes() ?>" id="panelBar">
     <?= $panelBar->pre() ?>
     <div class="panelBar-main" id="panelBar-main">
@@ -16,5 +10,6 @@ $panelBar = new Kirby\panelBar\Core([
   </div>
   <?= $panelBar->controls() ?>
   <?= $panelBar->post() ?>
-
+<?php else : ?>
+  <?= $panelBar->login() ?>
 <?php endif ?>